Changed hybrid index set builder methods so they take a non-const reference to a HybridISet instead of returning a pointer to an HybridISet. This make object ownership more clear.
Then, added HybridISet constructors to use the builder methods.  This allows
hybrid index sets to be created on the stack instead of forcing them to be
allocated on the heap.

Reorganized the "simple" test to run through more index set construction options
and more clearly print out pass/fail results.