iterator implemented for testing
This version uses an iterator-based interface for both RangeSegment and
ListSegment execution along with policy implementations for simd, serial
and the standard OpenMP options.  There are extra "range" functions in
each of the policies for now, and the iterator-based version does not
implement the Icount variants yet, but at least for kripke it looks all
good so far.  Lulesh is harder to tell, because the indexsets are still
implemented the original way, that said, that also seems promising so
far.