fixes for build on bgq
This alters the BGQ host_config files to use reasonably up-to-date c++
standard libraries.  The clang solution is not ideal, it really should
use the gcc-4.8.4 libstdc++ as well, but because of the way the wrapper
script is written, that's infeasible without completely reimplementing
the wrapper logic.  The libc++ version works fine for tests, but it's
conceivable it could cause issues for some code that uses raja in a
component and thus ends up with two standard libraries.