Add target forall and create target_reduce_openmp
Created a policy that adds OnDevice as a parameter and policies
omp_target_parallel_exec (same as parallel exec but with OnDevice=true)
omp_target_reduce

Added omp target statement to for all with parameter for OnDevice to be
able to turn on/off mapping to device easily

Created file target_reduce_openmp that holds instances of Reduce types that
use omp_target_reduce policy
ReduceSum is modeled after the non target version of ReduceSum. It holds a couple
of new data items that are used to do the sum on the device. The parent pointer cascade
back to the original host calling forall is avoided and the copy back is only done once.

Some of the other stuff in this commit is just me testing things to get around linking issues.