view and layout rework
Layout's type now depends only on the number of dimensions of the target.  Strong typing for views is now part of the view type, and permutations are either done as part of the type with a PermutedLayout or by generating an appropriately permuted base Layout with make_permuted_layout<PERM>(sizes...).