Add MultiPolicy wrapper policy for dynamic policy selection
This code still needs work, but it's enough to discuss the interface/implementation.  MultiPolicy stores a functor, or lambda or function pointer or whatever, which takes an iterable and based on that and whatever other data it stores internally decides which policy to invoke, and selects it by returning an integer offset into a list of policies in the parameter pack passed to MultiPolicy.  I'm not sure I love this interface, but it works and at least gives us a place to start.  The big questions are:

* What should the MultiPolicy selector take to choose a policy
* How does the selector signal the selected policy

The answer could even be that it takes instances of all of the available policies and directly invokes one, but I'm trying to avoid that if possible.  Perhaps ironically the MultiPolicy currently can't handle a nested MultiPolicy, because it only works with types rather than instances, but that should be easily fixable.