Minor perturbation to how we pull in GoogleTest.
Reworking build system for perf suite. Now RAJA is a submodule
and BLT is used to build it. BLT and RAJA each pull in their own
copy of GoogleTest which breaks compilation when RAJA testing is enabled.
This minor change allows the new system to work properly and doesn't
adversely affect the current state of RAJA.

If we switch RAJA over to use BLT as well, a lot of our CMake stuff
will go away.