the new nested::forall runs!
This version adds type computations to calculate the index tuple type,
an executor that handles For<index,policy,index_type> levels, and a
wrapper that carries loop data through levels of the process avoiding
copy construction of the loop body more than once as part of this
machinery.  In fact, work may need to be done that threaded reductions
continue to work inside this machinery.