Collapse policy construct added
The main addition is the extraction of sub-constructs from collapse to
generate the for list and allow a single executor policy to handle
multiple levels.  The mechanism used for this is extensible by deriving
a new clas from ForList and ensuring an as_for_list alias exists in the
type to provide the list of For-compatible constructs that it
represents.