types from segments, For refinement, strong typing
I didn't expect to do this this way, but in fixing an issue with
index_tuple type derivation to make the need for CFor go away this now
supports deriving the type of each index from the value_type of the
iterator type of the iterable at each index.  If a TypedFor is used,
that can override the type from the segment and re-construct the index
into the specified type, supporting the current strongly-typed index
system without modification while making it easier to transition to
strongly typed segments.  Also the strong typing is now working.