actually use the highest std flag we find, up nvcc
This just fixes some odd inconsistencies we had in applying c++ standard
selection arguments to the compiler.  Also adds support for the -std
c++14 argument to nvcc 9.  I've tested this with nvcc 9 on clang, and
gcc-4.9 separately and both work.  Will see if they work *together* on
travis.  May need an extra check for host compiler type and/or version
before selecting that nvcc flag.