fix xl openmp target backend
This PR fixes all build and run issues for OpenMP target offload with
the XL compiler as of the Feb 19 version.  Old workarounds removed,
atomic ref tests broken out to parallelize them and reduce amount of
them compiled by default, and mark a destructor declare target that is
not getting implicitly marked for some reason.