Consistent naming of generic configuration
Here I am doing two things:
- Renaming the folders so that they depend on $SYS_TYPE value and are
consistent with the ones in .gitlab/conf, but also with Umpire.
- Adding consitency in the naming of the files themselves. The idea
being that by replacing 'X' by '*' the file name can be used as a
pattern. This way, when choosing a config in .gitlab/conf/ we can
associate it with one of these generic config file (see following
commits).

Note:
As of today, there is only one generic config matching any of the
config files in .gitlab/conf. This is convenient, and shouldn’t change
without adapting GitLab build script accordingly (see following
commits).