Updating RAJA build scripts accordingly
Host-configs files have been split in anticipation of sharing part of
them with other projects. However this change, initially done for Gitlab
CI, has an impact on build scripts as well. The consequence is "dependency"
of build scripts on .gitlab directory. However, the configuration files
in .gitlab should move in a more generic place later on.