LC GitLab CI configuration
Now introducing the generic build script for GitLab CI, and the CI
configuration yaml files.

The aim is to have the simplest possible build workflow, with
configuration provided by environment variables (could be by arguments).
For now, the code to select the generic configuration is in the script,
but will be move in a bash module, to be reusable for testing purpose.