range*segment: clamp range for wrapping or inf.
Rather than either throwing (RangeSegment) or wrapping
(RangeStrideSegment) the ranges now clamp their begin to equal their end
in these corner cases.  This commit also includes some last tweaks for
TBB, and converts tests for the throw from RangeSegment to checks that
the begin and end are equivalent in this case.