Move zip_ref to separate zip_tuple header
Refactor zip_ref into zip_tuple.
Zip_tuple is designed to be used as if it was a reference
to its individual parts. This allows copy/move construction/assignment
to be done for each of its parts based on the reference semantics of
the zip_tuple.
Still need to figure out how to be able to make get<I>(zip_tuple)
work without conflicting with camp::get<I>(tuple).