switch workgroup test generatoin functions to macros
This fixes a weird issue with FindHIP that causes
unexpected errors from hipcc_cmake_linker_helper. This happens
because hip_add_executable overwrites a global variable which
does not work if functions are used as they scope the overwrite.
So the slightly invalid global default is used instead.
This default is set before HIP_CLANG_PATH is set so that path to
the compiler is omitted causing hipcc_cmake_linker_helper to attempt
to invoke the_first_argument/clang.