ensure tuple device copy constructor generates
The issue with camp tuple copy constructors seems to have come down to
the segment_tuple copy construction requested in LoopData's constructor.
Why on earth that would be I do not know, it also only appears to happen
on sierra-like systems for some reason.  Regardless, making the LoopData
constructor host/device takes care of it for the minimal reproducer.