ENH: use cmake_dependent_option with BLT options
The "ENABLE_TESTS", "ENABLE_EXAMPLES" and "ENABLE_BENCHMARKS"
are options defined in BLT. RAJA is re-defining and shadowing
these options. This generally works fine, however, when RAJA
is used as a sub-module within a super-project, this approach
limits the ability to control at a finer granularity which
project's tests, examples etc. are built. For example, it is
desirable to have the ability to disable RAJA's tests and
examples when building the parent project.

This commit changes the BLT defined options to corresponding
cmake_dependent_options, which gives super-projects more
fine-grain control.