Use cob and rocPRIM from install if available
Default to trying to find cub and rocPRIM in the cuda toolkit and rocm
install respectively when ENABLE_EXTERNAL_CUB and ENABLE_EXTERNEL_ROCPRIM
respecitively are Off. Add cub and rocm includes to PC_C_FLAGS. Allow
dependency on external version of cub or rocPRIM libraries even when
compiling with different backend to allow external cub to be used with
when hip is enabled.