Fix reset for cuda and hip reducers
Previously reset would call a base class reset expecting
two ValueLoc objects. The value and idx arguments to reset
were implicitly cast to ValueLoc objects and used as the value
and identity arguments to the base class reset. Thus an invalid
identity value was set causing subsequent uses to produce the wrong
result depending on the input values relation to the identity.