Add an unchecked iteration_mapping
This iteration mapping assumes that the number of
iterations is the same as the size of the range
and does no checking.
This is useful when mapping gpu blocks as we often
launch the exact number we need and don't need to check
if we are in range. This can give ~5% speedup vs direct
in this case.