Make global function params const
For cuda make them __grid_constant__ as well. Grid constant is a
special decorator for parameters that tells the compiler that you
won't modify the parameter values and allows for better optimizations.
Hip appears to assume this for all parameters.
The kinds of optimizations that nvcc can do with this are accessing
arrays with dynamic indices directly from the parameters isntead of
copying the array to local memory and dynamically accessing it there.

I did not do this for "forall params" params as reducers do modify
their value. We could work around this by making a copy in each
thread like we do for the old reducer interface.

It is possible for a user to use a functional object with mutable data
members similar to how the old reducer interface works which would break
this but for now it looks like we make copies of the loop body so that
should be ok.